+++ splineABS/NEWS.md
@@ -35,3 +35,4 @@
   Now, every subset in ABSpline shares the same basis. Copy basis into each subsets to to garentee this. The modified functions are adapted for this change.
   
 * When fitting model, each subset uses global basis and local data.
+* Allow input subsets instead of randomly genrating them. See function `gen.subset`.